Add 'pickup-line' command
This commit adds a new '!pickup-line' command to the bot. This command sends a random pickup line using the 'pickup-lines-api' API. The command requires an API token which is stored in the '.env' file and loaded using the 'dotenv' package. If an error occurs while fetching the pickup line, the bot will reply to the message that triggered the error with an error message.