Add 'inuslt' command
This commit adds a new command called 'insult', which sends a random insult using the 'evilinsult.com' API. The command can be triggered by typing '!insult' in a channel. The command will mention a user if one is specified, and will reply with an insult. The 'discord-reply' package is used to reply to the message that triggered the command in case an error occurs.