Update LICENSE to include full name in copyright notice
This commit updates the LICENSE.md file to include the full name of the copyright holder in the copyright notice. The previous version only included the first name 'Anton', but this version adds the full name 'Anton Van Assche'. The change was made to ensure that the copyright notice accurately reflects the copyright holder's full name.