Add 'pp-size' command
This commit adds a new command called 'pp-size' to the bot. The command generates a random number between 1 and 20, and creates a string of equal signs to represent the size of a pp. The length of the string is then padded with spaces to make it 20 characters long. The command also includes a humorous title based on the size of the pp and displays it in an embed.