Add 'pp-size' command to README
This commit adds information about the new '\!pp-size' command to the README file, along with an example usage. The command allows users to see the size of their or someone else's pp, and is activated by using the command followed by a mention of the desired user (e.g. '\!pp-size @bob').