Add '8ball' command to predict the future
This commit adds a new command, '8ball', to the bot. When a user asks a question using the command, the bot will randomly select an answer from a list of predefined responses and reply with an embedded message containing the answer. If an error occurs, the bot will reply to the message that triggered it with an error message.