Add 'lovemeter' command
This commit adds a new command 'lovemeter' which calculates the love percentage between two mentioned users and displays a fun message based on the result. The command validates that at least one user is mentioned and at most two users are mentioned, and if an error occurs, it will send an error message to the user who triggered it.