Improve command list formatting and add new commands
Previously, the command list was a simple bullet-point list. This commit
updates the formatting of the command list in the README.md file to use a
markdown table, which makes it easier to read and understand the different
commands that the bot can perform.

In addition, this commit adds several new commands to the list, including
'how-gay', 'lovemeter', and '8ball', as well as updating the 'pp-size' command
to show the size of a mentioned user's pp. The note at the bottom of the
README.md file has also been updated to remind users that they can change the
default prefix in the 'config.json' file.