Add 'compliment' command and update 'help' command
This commit adds a new 'compliment' command that sends a random compliment to the mentioned user or the first argument if no user is mentioned. The compliments are randomly selected from a list of titles and fetched compliments.

The help command has also been updated to include the new compliment command.